requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.0
pydantic==2.4.2
pytest==7.4.3
transformers==4.36.0
torch==2.6.0
